Make seperate library for resource parser and revert changes to core.
