+++ Microsoft.Health.Fhir.sln
@@ -150,6 +150,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Exporter", "tools\Exporter\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ResourceParser", "tools\ResourceParser\ResourceParser.csproj", "{86130CF7-35BA-45B3-8BE7-B83D382B32C3}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Health.Fhir.R4.ResourceParser", "tools\Microsoft.Health.Fhir.R4.ResourceParser\Microsoft.Health.Fhir.R4.ResourceParser.csproj", "{2CAC8D57-180F-4DD4-B11D-6B2AB65B2A48}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -328,6 +330,10 @@ Global
 		{86130CF7-35BA-45B3-8BE7-B83D382B32C3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{86130CF7-35BA-45B3-8BE7-B83D382B32C3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{86130CF7-35BA-45B3-8BE7-B83D382B32C3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2CAC8D57-180F-4DD4-B11D-6B2AB65B2A48}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2CAC8D57-180F-4DD4-B11D-6B2AB65B2A48}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2CAC8D57-180F-4DD4-B11D-6B2AB65B2A48}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2CAC8D57-180F-4DD4-B11D-6B2AB65B2A48}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -397,10 +403,11 @@ Global
 		{F6B94905-B496-46AD-B2A7-2ABCB0B2A6B4} = {B70945F4-01A6-4351-955B-C4A2943B5E3B}
 		{E468B6C6-9098-4293-AFD6-3B1675D67063} = {B70945F4-01A6-4351-955B-C4A2943B5E3B}
 		{86130CF7-35BA-45B3-8BE7-B83D382B32C3} = {B70945F4-01A6-4351-955B-C4A2943B5E3B}
+		{2CAC8D57-180F-4DD4-B11D-6B2AB65B2A48} = {B70945F4-01A6-4351-955B-C4A2943B5E3B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		RESX_SortFileContentOnSave = True
 		SolutionGuid = {E370FB31-CF95-47D1-B1E1-863A77973FF8}
+		RESX_SortFileContentOnSave = True
 	EndGlobalSection
 	GlobalSection(SharedMSBuildProjectFiles) = preSolution
 		test\Microsoft.Health.Fhir.Shared.Tests.E2E.Common\Microsoft.Health.Fhir.Shared.Tests.E2E.Common.projitems*{0478b687-7105-40f6-a2dc-81057890e944}*SharedItemsImports = 13
