+++ src/Microsoft.Health.Fhir.Core/Properties/AssemblyInfo.cs
@@ -41,3 +41,5 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Microsoft.Health.Fhir.R4.Tests.E2E")]
 [assembly: InternalsVisibleTo("Microsoft.Health.Fhir.R5.Tests.E2E")]
 [assembly: InternalsVisibleTo("Microsoft.Health.Fhir.Stu3.Tests.E2E")]
+
+[assembly: InternalsVisibleTo("Microsoft.Health.Fhir.R4.ResourceParser")]
